feature(landing page): convert to approved file structure
break down page into smaller components
place component chunks in appropriate file structures
set up tests for components
