+++ package.json
@@ -4,7 +4,7 @@
   "description": "A Social platform for the creative at heart",
   "main": "index.js",
   "private": true,
-  "proxy": "http://localhost:3200",
+  "proxy": "http://localhost:5000",
   "dependencies": {
     "axios": "^0.18.0",
     "bootstrap": "^4.1.3",
@@ -27,7 +27,6 @@
     "eject": "react-scripts eject",
     "lint": "node_modules/.bin/eslint --fix ./"
   },
-  "proxy": "http://localhost:5000",
   "eslintConfig": {
     "extends": "react-app"
   },
