+++ src/containers/ArticleCard.js
@@ -1,18 +1,28 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import getArticles from '../actions/articleActions';
+import Loading from '../components/shared/Loading';
 
 class Card extends Component {
+  componentDidMount() {
+    const { getArticles: getAllArticles } = this.props;
+    getAllArticles();
+  }
+
   state = {
     articleEnd: 6,
   };
 
   sortArticles = (articles) => {
-    const sortedArticles = articles.map((each) => {
-      const filteredArticles = this.filterLikes(each.reactions);
-      return { article: each, track: filteredArticles };
-    }).sort((a, b) => {
-      const result = b.track - a.track;
-      return result;
-    });
+    const sortedArticles = articles
+      .map((each) => {
+        const filteredArticles = this.filterLikes(each.reactions);
+        return { article: each, track: filteredArticles };
+      })
+      .sort((a, b) => {
+        const result = b.track - a.track;
+        return result;
+      });
     return sortedArticles;
   };
 
@@ -43,66 +53,74 @@ class Card extends Component {
 
   displayCards = (displayArticles) => {
     const { articleEnd } = this.state;
-    return displayArticles(articleEnd).map(
-      (card) => {
-        const displayedBody = card.article.body.slice(0, 120);
-        const displayedTitle = card.article.title.slice(0, 30);
-        return (
-          <div
-            onClick={this.getArticle}
-            className="hero-card"
-            key={card.article.id}
-          >
-            <div className="hero-card-details col-sm-7">
-              <h6>{displayedTitle}</h6>
+    return displayArticles(articleEnd).map((card) => {
+      const displayedBody = card.article.body.slice(0, 120);
+      const displayedTitle = card.article.title.slice(0, 30);
+      return (
+        <div
+          onClick={this.getArticle}
+          className="hero-card"
+          key={card.article.id}
+        >
+          <div className="hero-card-details col-sm-7">
+            <h6>{displayedTitle}</h6>
+            <p>
+              {displayedBody}
+              ...
+            </p>
+            <div className="hero-card-author">
+              {card.article.user.profile.image === null ? (
+                <img
+                  src="https://res.cloudinary.com/drmmqcxkc/image/upload/v1541581893/Authors%20Haven/user-placeholder.png"
+                  alt="Author Profile"
+                />
+              ) : (
+                <img
+                  src={card.article.user.profile.image}
+                  alt="Author Profile"
+                />
+              )}
               <p>
-                {displayedBody}
-                ...
+                {card.article.user.firstname} {card.article.user.lastname}
               </p>
-              <div className="hero-card-author">
-                {card.article.user.profile.image === null ? (
-                  <img
-                    src="https://res.cloudinary.com/drmmqcxkc/image/upload/v1541581893/Authors%20Haven/user-placeholder.png"
-                    alt="Author Profile"
-                  />
-                ) : (
-                  <img src={card.article.user.profile.image} alt="Author Profile" />
-                )}
-                <p>
-                  {card.article.user.firstname} {card.article.user.lastname}
-                </p>
-              </div>
             </div>
-            {card.article.image === null ? (
-              <div className="hero-card-image col-sm-5">
-                <img
-                  src="https://res.cloudinary.com/drmmqcxkc/image/upload/v1541426068/Authors%20Haven/icons8-heart-outline-48-grey.png"
-                  alt=""
-                />
-                <p>{this.filterLikes(card.article.reactions)}</p>
-              </div>
-            ) : (
-              <div
-                style={this.changeBackground(card.article.image)}
-                className="hero-card-image col-sm-5"
-              >
-                <img
-                  src="https://res.cloudinary.com/drmmqcxkc/image/upload/v1541426068/Authors%20Haven/icons8-heart-outline-48-grey.png"
-                  alt=""
-                />
-                <p>{this.filterLikes(card.article.reactions)}</p>
-              </div>
-            )}
           </div>
-        );
-      },
-    );
-  }
+          {card.article.image === null ? (
+            <div className="hero-card-image col-sm-5">
+              <img
+                src="https://res.cloudinary.com/drmmqcxkc/image/upload/v1541426068/Authors%20Haven/icons8-heart-outline-48-grey.png"
+                alt=""
+              />
+              <p>{this.filterLikes(card.article.reactions)}</p>
+            </div>
+          ) : (
+            <div
+              style={this.changeBackground(card.article.image)}
+              className="hero-card-image col-sm-5"
+            >
+              <img
+                src="https://res.cloudinary.com/drmmqcxkc/image/upload/v1541426068/Authors%20Haven/icons8-heart-outline-48-grey.png"
+                alt=""
+              />
+              <p>{this.filterLikes(card.article.reactions)}</p>
+            </div>
+          )}
+        </div>
+      );
+    });
+  };
 
   render() {
+    const { articles } = this.props;
     return (
       <div>
-        <div className="card-wrapper">{this.displayCards(this.displayArticles)}</div>
+        {articles ? (
+          <div className="card-wrapper">
+            {this.displayCards(this.displayArticles)}
+          </div>
+        ) : (
+          <Loading />
+        )}
         <div onClick={this.handleClick} className="hero-moreArticles row">
           <p>More Articles</p>
           <img
@@ -115,4 +133,8 @@ class Card extends Component {
   }
 }
 
-export default Card;
+function mapStatetoProps(state) {
+  return { articles: state.articles.articles.articles, error: state.articles.error.response };
+}
+
+export default connect(mapStatetoProps, { getArticles })(Card);
