+++ src/styles/components/landingPage.scss
@@ -254,10 +254,17 @@ $faintblue: #F4F9FC;
     .row{
       .hero-article{
       margin: 50px;
-      @media only screen and (max-width: 900px) {
-        margin: 17px;
-      }
+        @media only screen and (max-width: 900px) {
+          margin: 17px;
+        }
       }
     }
+  }
+}
+
+.hero-body-background{
+  background-color: white;
+  @media only screen and (max-width: 900px) {
+    margin-top: -70px;
   }
 }
