+++ src/tests/containers/App.test.js
@@ -13,7 +13,7 @@ describe('App UI', () => {
   describe('render features', () => {
     test('container should render as expected', () => {
       const tree = toJson(wrapped);
-        expect(tree).toMatchSnapshot();
-      });
+      expect(tree).toMatchSnapshot();
     });
+  });
 });
