+++ src/tests/containers/Signup.test.js
@@ -20,12 +20,12 @@ describe('Signup UI', () => {
   describe('render features', () => {
     test('container should render as expected', () => {
       const tree = toJson(wrapped);
-        expect(tree).toMatchSnapshot();
-      });
-    
-      it('has two input fields and a button', () => {
-        expect(wrapped.find('input').length).toEqual(5);
-        expect(wrapped.find('button').length).toEqual(1);
-      });
+      expect(tree).toMatchSnapshot();
     });
+
+    it('has two input fields and a button', () => {
+      expect(wrapped.find('input').length).toEqual(5);
+      expect(wrapped.find('button').length).toEqual(1);
+    });
+  });
 });
