--- src/tests/sample.test.js
@@ -1,15 +1,0 @@
-import authReducer from '../reducers/authReducer';
-import { registerUser } from '../actions/authActions';
-
-describe('auth reducer', () => {
-  test('is correct', () => {
-    const finalState = {
-      articles: [],
-      error: [],
-      isAuthenticated: false,
-      user: { name: 'tony', email: 'tony@gmail.com' },
-    };
-    const action = registerUser({ name: 'tony', email: 'tony@gmail.com' });
-    expect(authReducer(undefined, action)).toEqual(finalState);
-  });
-});
