+++ src/tests/views/LandingPage.test.js
@@ -1,17 +1,27 @@
 import React from 'react';
-import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
+import { mount } from 'enzyme';
+import { MemoryRouter } from 'react-router-dom';
+import Root from '../../root';
 import LandingPage from '../../views/LandingPage';
+import '../../styles/styles.scss';
 
 let wrapped;
-
 beforeEach(() => {
-  wrapped = shallow(<LandingPage />);
+  wrapped = mount(
+    <Root>
+      <MemoryRouter initialEntries={[{ key: 'testkey' }]}>
+        <LandingPage />
+      </MemoryRouter>
+    </Root>,
+  );
 });
 
-describe('LandingPage UI', () => {
+afterEach(() => wrapped.unmount());
+
+describe('Landing Page UI', () => {
   describe('render features', () => {
-    test('view should render as expected', () => {
+    test('container should render as expected', () => {
       const tree = toJson(wrapped);
       expect(tree).toMatchSnapshot();
     });
